feat: add basic routing and comparison state
