+++ apps/web/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router-dom": "^7.6.3",
     "zustand": "^5.0.6"
   },
   "devDependencies": {
