+++ apps/web/src/store/jobStore.ts
@@ -13,10 +13,17 @@ interface JobState {
   history: JobSummary[];
   metrics?: Record<string, number>;
   running: boolean;
+  comparison: {
+    baseJobId?: string;
+    compareJobId?: string;
+  };
   start(job: JobSummary): void;
   append(text: string): void;
   finish(metrics: Record<string, number>): void;
   reset(): void;
+  setBaseJob(id: string): void;
+  setCompareJob(id: string): void;
+  clearComparison(): void;
   loadHistory(): Promise<void>;
 }
 
@@ -24,6 +31,7 @@ export const useJobStore = create<JobState>((set) => ({
   log: [],
   history: [],
   running: false,
+  comparison: {},
   start: (job) => {
     console.log('🏁 Store: Starting job', job);
     set({ current: job, log: [], metrics: undefined, running: true });
@@ -39,6 +47,15 @@ export const useJobStore = create<JobState>((set) => ({
   reset: () => {
     console.log('🔄 Store: Resetting');
     set({ current: undefined, log: [], metrics: undefined, running: false });
+  },
+  setBaseJob: (id) => {
+    set((s) => ({ comparison: { ...s.comparison, baseJobId: id } }));
+  },
+  setCompareJob: (id) => {
+    set((s) => ({ comparison: { ...s.comparison, compareJobId: id } }));
+  },
+  clearComparison: () => {
+    set(() => ({ comparison: {} }));
   },
   loadHistory: async () => {
     console.log('📜 Store: Loading history');
